Refactor. En proceso trasladar reses.

+++ GestionGanadera/src/animation/Animation.java
@@ -5,6 +5,7 @@ import java.awt.image.BufferedImage;
 
 /**
  * This class represents the animation of puts sprite sheets in a sequence.
+ * 
  * @author Juan Camilo Vélez Olaya
  *
  */
@@ -14,37 +15,37 @@ public class Animation {
 	 * The speed of the animation
 	 */
 	private int speed;
-	
+
 	/**
 	 * The quantity of frames (Sprite Sheets)
 	 */
 	private int frames;
-	
+
 	/**
 	 * An auxiliar index
 	 */
 	private int index = 0;
-	
+
 	/**
 	 * An auxiliar index to next frames method.
 	 */
 	private int count = 0;
-	
+
 	/**
 	 * The number of columns of the image to run the animation.
 	 */
 	private int col;
-	
+
 	/**
 	 * The number of rows of the image to run the animation
 	 */
 	private int row;
-	
+
 	/**
 	 * 
 	 */
 	private int startRow;
-	
+
 	/**
 	 * 
 	 */
@@ -54,17 +55,17 @@ public class Animation {
 	 * It is the image to get the sprite sheets.
 	 */
 	private BufferedImage image;
-	
+
 	/**
 	 * It is an array of the sub images (sprites)
 	 */
 	private BufferedImage[] images;
-	
+
 	/**
 	 * It is the current sprite for the animation.
 	 */
 	private BufferedImage currentImage;
-	
+
 	/**
 	 * Variable to know if the animation has finished.
 	 */
@@ -72,11 +73,12 @@ public class Animation {
 
 	/**
 	 * Constructor to create an animation
-	 * @param image The image with sprites.
+	 * 
+	 * @param image  The image with sprites.
 	 * @param frames The quantity of sprites.
-	 * @param speed The speed of the animation
-	 * @param col The number of columns that the image has of sprites.
-	 * @param row The number of rows that the image has of sprites.
+	 * @param speed  The speed of the animation
+	 * @param col    The number of columns that the image has of sprites.
+	 * @param row    The number of rows that the image has of sprites.
 	 */
 	public Animation(BufferedImage image, int frames, int speed, int col, int row) {
 
@@ -93,7 +95,7 @@ public class Animation {
 		fillSprites();
 
 	}
-	
+
 	/**
 	 * 
 	 * @param image
@@ -102,7 +104,7 @@ public class Animation {
 	 * @param col
 	 * @param row
 	 */
-	public Animation(BufferedImage image, int frames, int speed,int startCol, int startRow, int col, int row) {
+	public Animation(BufferedImage image, int frames, int speed, int startCol, int startRow, int col, int row) {
 
 		this.image = image;
 		this.startCol = startCol;
@@ -118,7 +120,6 @@ public class Animation {
 
 	}
 
-
 	/**
 	 * Puts all the sprites sheets in an array.
 	 * 
@@ -133,7 +134,7 @@ public class Animation {
 
 			for (int j = startCol; j <= col; j++, k++) {
 
-				images[k] = ss.grabImage(i, j, (int) image.getWidth() / col, (int) image.getHeight() / row);
+				images[k] = ss.grabImage(i, j, image.getWidth() / col, image.getHeight() / row);
 
 			}
 
@@ -142,7 +143,7 @@ public class Animation {
 	}
 
 	/**
-	 * Change the current image with the next sprite. 
+	 * Change the current image with the next sprite.
 	 */
 	public void nextFrame() {
 
@@ -150,7 +151,8 @@ public class Animation {
 
 		count++;
 
-		if (count >= frames - 1) {  // You can modify frames - 1 to another number if your image does not have sprite sheets in all the image.
+		if (count >= frames - 1) { // You can modify frames - 1 to another number if your image does not have
+									// sprite sheets in all the image.
 									// If your image has sprite sheets in all the image you can put only frames.
 			alive = false;
 			count = 0;
@@ -168,7 +170,8 @@ public class Animation {
 	}
 
 	/**
-	 * Uses the method next frames to simulate an animation changing frames in an execution
+	 * Uses the method next frames to simulate an animation changing frames in an
+	 * execution
 	 */
 	public void runAnimation() {
 
@@ -184,10 +187,11 @@ public class Animation {
 
 	/**
 	 * Draws the animation
-	 * @param g The graphics to draw.
-	 * @param x Pos in X
-	 * @param y Pos in Y
-	 * @param offset Move the position in X 
+	 * 
+	 * @param g      The graphics to draw.
+	 * @param x      Pos in X
+	 * @param y      Pos in Y
+	 * @param offset Move the position in X
 	 */
 	public void drawAnimation(Graphics g, double x, double y, int offset) {
 
@@ -196,26 +200,26 @@ public class Animation {
 	}
 
 	public void setImage(BufferedImage image) {
-		
-	this.image = image;	
-		
+
+		this.image = image;
+
 	}
-	
+
 	public void setFrames(int frames) {
-	
-	this.frames = frames;
+
+		this.frames = frames;
 	}
-	
+
 	public void setCol(int col) {
-	
+
 		this.col = col;
-		
+
 	}
-	
+
 	public void setRow(int row) {
-		
-	this.row = row;	
-		
+
+		this.row = row;
+
 	}
-		
+
 }
